Fixed a major bug for sbsdig jobs..
I wasn't aware that LIBSBSDIG has dependence on G4SBSRunData class for proper execution. In the run-sbsdig.sh I only source sbsdigenv.(c)sh which doesn't make G4SBSRunData class available in the ROOT session. This makes HCAL DB all screwed up. It is by all means should be a fatal error but the software as it is only throws some warning at the time of digitization and finishes the job. This needs to be changed! Anyways, for now I am sourcing g4sbs.csh in the run-sbsdig.sh script as well, which should solve the issue.