Implemented module system load general libraries (and/or apps)..
Until now we were using /site/12gev_phys/softenv.sh to load apps like root, gcc, python etc. It worked but the whole system is very unstable and it forced us to use a non-standard analyzer build. The much better option is to use /group/halla/modulefiles that are very stable and maintained by Ole. It also provide a stable build of analyzer and all the associated libraries.

I have implemented it in such a way that one can switch between the above mentioned two methods of setting environments very easily. Everything is controlled by `setenv.sh` file. Here is a quick how-to:
1. `ANAVER` sets the version of analyzer build to load from modulefiles. Eg. ANAVER='1.7.4' prompts `module load analyzer/1.7.4`
2. `JLABENV` sets the version of 12gev_phys/softenv.sh to choose. Eg. JLABENV='2.6' prompts `source /site/12gev_phys/softenv.sh 2.5`
3. `useJLABENV` flag can be used to switch between `module` and `12gev_phys` system.
   - `useJLABENV=1` (ie True): Forces all the `run-` scripts to switch from `module` to `12gev_phys` environments.
   - `useJLABENV!=1` (ie False): Defaults to `module` system.

**NOTE: Above variables affects all the `run-` scripts similarly for the sake of standardization. User is expected to build any additional library (e.g. simc, g4sbs, libsbsdig, etc) with the same versions of the above mentioned softwares (ie analyzer, root, gcc, etc.).